requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
fastapi==0.110.2
uvicorn==0.29.0
requests==2.31.0
pydantic==2.7.1
cohere==5.3.4
transformers==4.40.1
python-dotenv==1.0.1